+++ commands/attack/update.ts
@@ -1,0 +1,65 @@
+import type { Subcommand } from '../../util/commands';
+import { ChannelType, EmbedBuilder, SlashCommandSubcommandBuilder } from 'discord.js';
+import { ATTACK_FORUM_CHANNEL_ID } from '../../config';
+
+// Utils
+import { lock, writePortsFile } from '../../modules/slack';
+import { execAsync } from '../../util/exec';
+import { updateInfoForTeam } from '../../bot';
+
+
+export default {
+    data: new SlashCommandSubcommandBuilder()
+        .setName('update')
+        .setDescription('Update the IP and port for the specified target.')
+        .addStringOption((option) => option
+            .setName('target')
+            .setDescription('The target to update.')
+            .setAutocomplete(true)
+            .setRequired(true))
+        .addStringOption((option) => option
+            .setName('ip')
+            .setDescription('The new IP.')
+            .setRequired(true))
+        .addIntegerOption((option) => option
+            .setName('port_low')
+            .setDescription('The new low port.')
+            .setRequired(true))
+        .addIntegerOption((option) => option
+            .setName('port_high')
+            .setDescription('The new high port.')
+            .setRequired(true)),
+
+    async execute(interaction) {
+        const target = interaction.options.getString('target', true);
+        const ip = interaction.options.getString('ip', true);
+        const portLow = interaction.options.getInteger('port_low', true);
+        const portHigh = interaction.options.getInteger('port_high', true);
+
+        await updateInfoForTeam(target, ip, portLow, portHigh);
+
+        await writePortsFile(target, ip, portLow, portHigh);
+        await lock.acquire('git', async () => {
+            await execAsync(`cd temp && git pull --ff-only && git add "${target}/" && git -c user.name="eCTF scrape bot" -c user.email="purdue@ectf.fake" commit -m "Update ports for ${target}" && git push`);
+        });
+
+        const successEmbed = new EmbedBuilder()
+            .setDescription('Successfully updated target info.')
+            .setColor('#C61130')
+        await interaction.reply({ embeds: [successEmbed] });
+    },
+    async autocomplete(interaction) {
+        const input = interaction.options.getFocused();
+
+        const attackThreadsChannel = interaction.client.channels.cache.get(ATTACK_FORUM_CHANNEL_ID);
+        if (attackThreadsChannel?.type !== ChannelType.GuildForum)
+            return interaction.respond([]);
+
+        const targets = attackThreadsChannel.threads.cache
+            .filter((c) => c.name.toLowerCase().includes(input.toLowerCase()))
+            .map((c) => ({ name: c.name, value: c.name }))
+            .slice(0, 25)
+
+        return interaction.respond(targets);
+    }
+} satisfies Subcommand;
